Add makefile skeleton
This makefile creates a dependency for tpm2lib, which expects all .c
files in this directory to be compiled and included in the library.

Setting variable obj in the command line or in Bash environment allows
to override the output directory, which is set to ./build by default.

None of the source files is compileable yet,

BUG=none
TEST=none yet

Change-Id: I0dff9bc3e39478e4cd9591f153a1712215ec2364
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>