Changes to allow compilation of CpriECC.c
Some openssl includes do not meet this module's expectations, a
definition of ec_group_st (which is tyoedefed to EC_CURVE) was added
to resolve compilation issues. This structure contents will have to be
refined.

CpriDataECC.c is not supposed to be a separate compilation unit, and
as such it is being removed from Makefile.

BUG=none
TEST=compilation succeeds:
   cc -Wall -Werror -c -o /dev/null CpriECC.c

Change-Id: Ib37e34d567f95af20a064acd70c706a6359edd8c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/274099
Reviewed-by: Darren Krahn <dkrahn@chromium.org>