build: fix {INT}_TO_ARRAY macros
Macros used to convert different size integers into arrays return
integers themselves using the comma operator. GCC throws a 'value
computed but not ussed' in this case, which can be pacified by
appending (void) in front of the macro invocation. For some reason
different invocations trigger GCC warnings inside and outside chroot.

But the bottom line is that the code does not used the values returned
by the macros, so a better fix is to change the macros not to return
the values.

BUG=none
TEST=make succeeds for x86 both inside and outside chroot now.

Change-Id: I8105b8051aee2cb72409dd70c5e5cfa4f841b549
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/275394
Reviewed-by: Bill Richardson <wfrichar@chromium.org>