improve _cpri__GenerateRandom stub
This function is supposed to return the number of generated random
characters. While we still are not generating any numbers, let's just
return the requested number, whoever calls this function will use
whatever was in the buffer when the function was called.

CR50 TRNG circuit is not quite functional yet, eventually it will be
used to feed true random numbers to the caller.

BUG=chrome-os-partner:43025
TEST=with this and some EC code changes trunksd successfully
     initializes CR50.

Change-Id: I5fe9354d0e37deeb54499bae6a7a28d77343c07b
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/293577
Reviewed-by: Bill Richardson <wfrichar@chromium.org>