generator: re-sync the scripts and source code
It turns out that the source files generated by running generator.py
are different from those checked in in the repository.

This patch tweaks the script to make generated files exactly match the
repository contents. In fact many of these changes are not needed, but
it is worth submitting this to provide a way to reproduce the existing
state of the source tree.

The README file is being modified to include instructions on how to
actually re-generate the source code.

BUG=chromium:501639
TEST=running the following command results in the generated files
     exactly matching the git contents:

$ generator/generator.py generator/structures.txt \
        generator/commands.txt

Change-Id: I3ff3c081108ba4cb7566ae53f840da5583f22c01
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295938
Reviewed-by: Darren Krahn <dkrahn@chromium.org>