upstream: Global variables in NVMem.c should be local
This is a change to the code scraped from the TCG spec.

The original code puts some external variables in PlatformData.c
(declared in PlatformData.h). The comments suggest that they're
used by Microsoft's simulator, but as far as the reference code
is concerned they are used only in NVMem.c.

If we later discover a need for similar variables, we can add
them back, along with reference code that uses them.

BUG=chrome-os-partner:43791
BRANCH=none,upstream
TEST=manual

  sudo emerge tpm2 tpm2-simulator

  cd src/platform/ec
  make buildall

Change-Id: Ia82bbd8ad8229e276a70dac5f944796fe1bd6468
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/296991
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>