Add NVMem.c to the list of HOST_SOURCES
NVMem.c implements file-backed non-volatile memory unless we're
compiling in EMBEDDED_MODE, in which case it does nothing.

Since the embedded system will have to implement the _plat__NV*
functions natively (not here), we might as well move this into
the host-only sources.

BUG=chrome-os-partner:44745
BRANCH=none
CQ-DEPEND=CL:296525
TEST=sudo emerge tpm2 tpm2-simulator

And, with the CQ-DEPEND applied:

  cd src/platform/ec
  make buildall

Change-Id: I90cf4818ab891555151aa1a5ee8afacce6f3086b
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/296476
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>