generator: modify the scripts to process HTML sources
The scripts scraping TPM2 specification are being changed to use the
specification directly (in HTML forms). The existing modules
(command_generator and structure_generator) are being modified to
provide the 'back end processing' for the data retrieved by the new
scraper (extract_structures and tpm_table).

The test is modified to still pass, test_structure_generator.html is
the source of test data.

README includes instructions how to re-generate the source code.

BUG=chromium:501639
TEST=new code generation works fine, the test also passes:
   $ ./generator/structure_generator_test.py
   .......
   ----------------------------------------------------------------------
   Ran 7 tests in 0.007s

   OK

Change-Id: I2a0e925475d4cd2b57528bfe1b162245fc6e0f43
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/300662