generator: re-generate source files
This patch is the result of running the code generation scripts from
the previous patch.

The main deviation from the code which existed earlier is the fact
that the TPM types are now generated (before they were manually copied
from an open source implementation).

TPM_Types.h had to be edited manually to exclude all copied stuff and
include the generated file (tmp_types.h).

The changes are minimal - out of the 223 generated files only a
handful differs, ans most of the changes are due to the actual
structure layout and set of supported algorithms changes.

BUG=chromium:501639
TEST=the only change is building the library standalone and as a part
     of the cr50 image, both succeed.

Change-Id: I4ef87a08c2457524d7e22d53a58964056414174c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/300663