generator: honor conditional field types
Some fields in the TMP objects are 'conditional', i.e. can be either
present or not in the marshaled data. Those fields are marked with the
trailing plus sign in the TCG TPM2 library specification Part 2. This
patch makes sure that the generated code accounts for these suffixes,
when present.

BUG=chrome-os-partner:44608
TEST=the list of 69 test fixed by this patch can be found in
      https://goo.gl/KIxxyM

Change-Id: I0fddcad867333bd13e03b4ab12d2ff5e9898cc26
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/305490
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>