generator: do not mix ECC and RSA types
Some tables in TCG TPM2 Specification Part 2 use so called algorithm
macros, which allow specify subsets of algorithms included in Table 9.
In certain cases the macros expand to a wider than necessary subset,
this is where the basic algorithm dependency included in the table
header in curly brackets ({ECC} or {RSA}) is coming in.

The dependency included in the table header must match the value in
the fourth column in table 9.

BUG=chrome-os-partner:44608
TEST=verified that generated code uses updated algorithm subsets

Change-Id: I7b007ead8888df80a208c4bb0949a85173427875
Reviewed-on: https://chromium-review.googlesource.com/305492
Commit-Ready: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>