Assign values to return code modifiers.
The library routines scraped from Part 4 of the TCG TPM2 spec often
add constant offsets to return codes. The values are not defined in
any of the spec's four parts, so they were included in TPM_Types.h for
future consideration.

The actual values of these offset codes are supposed to be calculated
at run time depending on status of the device and results of command
execution as described in "Table 15 - Format-One Response Codes" of
the Part 2 of the specification.

The plan is to properly integrate calculation of these values into the
code later, this patch is a catch all submission to make the tests
pass even if without fully informative return codes.

The Part 4 sources are also interchangeably using return codes
prefixed with TMP_RC_ and TPM_RCS_. It looks like they are identical,
so this patch just makes sure that the values with different prefixes
match.

BUG=chrome-os-partner:44608, chromium:544181
TEST=this patch fixes 21 tests, the list can be found at
     https://goo.gl/o6LDiR

Change-Id: I24f65a4508c84550f79c279b03f265ebca30759e
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/305387
Reviewed-by: Darren Krahn <dkrahn@chromium.org>