generator: do not check parameter values when unmarshaling
The specification defines different return error values for various
command parameter errors. While the first three fields of the header
need to be checked and rejected by the unmarshaling code, the errors
in the command code and algorithm ID parameters need to be processed
by the respective commands, generated error values are sometimes
command specific.

This patch modifies the scripts to generate the code where
unmarshsling errors in certain fields are ignored, so that the command
processing can take place and return the proper return value.

BUG=chrome-os-partner:44608
TEST=this change fixes two tests listed in https://goo.gl/ziOdWS

Change-Id: I47128165dc516a2d9dcad752f73f8f2f9d5335c8
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/305496
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>