generator: avoid nested invocations for marshaling/unmarshaling
A lot of TPM objects are in fact simple integers one, two or four
bytes in size. The TPM2 library spec defines some fairly long type
chains which in the end of the day are just simple integers. The
currently generated marshaling/unmarshaling code follows these chains
causing multiple nested invocations, which is both a waste of time and
stack space.

This patch modifies generated code to shortcut to the lowest base type
when marshaling/unmarshaling integer base objects.

BUG=none
TEST=the generated code passes the same set of tests as before this
     change.

Change-Id: I2c5028cd32f5d74f6fa6ab9919d5c6e0c5b9fea4
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/305599
Reviewed-by: Darren Krahn <dkrahn@chromium.org>