Replace TPM_RCS_xxx with TPM_RC_xxx
The code scraped from the TPM2 library uses return code names prefixed
with TPM_RC_. But in a few places in the code the TPM_RCS_ prefix is
used. TPM_RCS_ prefixed values are not defined anywhere, form the code
logic they are equivalent to the corresponding TPM_RC_ prefixed
values.

This patch makes sure the consistent naming is used.

BUG=chrome-os-partner:44608
TEST=the tests still pass.

Change-Id: I342de9bdc6dd1e70e3c8268d38f02c9429980cfc
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/306650
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>