Move stubs.c to the package where it is implemented
When running in embedded mode, the stubs.c file is supposed to provide
implementation of numerous library functions not available on the
device from standard libraries.

The implementation is extremely device specific, and is likely to be
different for different devices, this is why stubs.c should be placed
in the implementation tree, not in the common TPM2 tree.

CQ-DEPEND=CL:307043
BUG=chrome-os-partner:43025
TEST=the tree still builds after the file has been moved

Change-Id: Id658389197baa0f0ef70d5356076a8df5552ee36
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/306709
Reviewed-by: Bill Richardson <wfrichar@chromium.org>