generator: report error if input buffer is larger than expected
One of the tests expects the TPM to return the TPM_RC_SIZE error in
case the buffer containing the digest exceeds the expected size of the
digest, but the simulator ignores this size discrepancy.

Let's make the implementation exact and always return TPM_RC_SIZE if
unmarshaling returns success but there are some leftover data.

BUG=chrome-os-partner:43025
TEST=fixed test is described in https://goo.gl/kQnud0

Change-Id: Ifcd98b09070b38356ee7975bff3b1710d34e17b9
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/307601
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>