Fix RSA_Decode OAEP fill handling
OAEP expects that there would be at least two bytes filling between
the message and the padding, all but the last filling bytes set to
zero, and the last byte s set to 1. The TPM is expected to reject OAEP
RSA encoded messages if this convention is not followed.

This patch enforces this behavior (any other than 0x1 terminating byte
is rejected).

BUG=chrome-os-partner:43025
TEST=the fixed test is described https://goo.gl/lE8ma3

Change-Id: Iac1dcd596a6048328818cf8aa131e9b8e7735d92
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/309166
Commit-Ready: Vadim Bendebury <vbendeb@google.com>
Tested-by: Vadim Bendebury <vbendeb@google.com>
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>