Add optional prefix to library object names
The caller of the makefile may specify OBJ_PREFIX to prefix all object
filenames in the archive with a common string.  This allows the caller's
linker script to group the data and bss sections for this library in one
place.  For example, if OBJ_PREFIX=Tpm2_, the caller's linker script can
have something like

	__bss_libtpm2_start = .;
	Tpm2_*(.bss)
	__bss_libtpm2_end = .;

Using a unique prefix is necessary in this case because files in
archives only have a filename, not a full path.

If OBJ_PREFIX is not specified, the makefile works as it did before.
That is, this is backwards compatible with any other users.

BUG=chrome-os-partner:52366
BRANCH=none
TEST=make
     ar t build/libtpm2.a --> object files match source files
     make clean
     make OBJ_PREFIX=foo_
     ar t build/libtpm2.a --> object files are prefixed with foo_
     make clean

Change-Id: Ifa16545b6dd765874a54268055db10a37a29fafc
Signed-off-by: Randall Spangler <rspangler@chromium.org>
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/366792